refactor: move TextEditor to `components/ui`
